Conflicts:
	scilab/CHANGES_5.4.X
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln
	scilab/Visual-Studio-settings/ant-all/ant-all.vcxproj
	scilab/modules/commons/Makefile.am
	scilab/modules/commons/Makefile.in
	scilab/modules/commons/src/c/commons.vcxproj
	scilab/modules/commons/src/c/commons.vcxproj.filters
	scilab/modules/core/sci_gateway/c/sci_exit.c
	scilab/modules/fileio/fileio.vcxproj.filters
	scilab/modules/gui/src/jni/CallScilabBridge.cpp
	scilab/modules/gui/src/jni/CallScilabBridge.hxx
	scilab/modules/history_manager/sci_gateway/c/sci_historymanager.c
	scilab/modules/scinotes/src/java/org/scilab/modules/scinotes/EditorComponent.java
	scilab/modules/scinotes/src/java/org/scilab/modules/scinotes/ScilabLexer.java
	scilab/modules/shell/src/c/GetCommandLine.c
	scilab/modules/shell/src/c/libjvm_Import.def
	scilab/modules/ui_data/src/c/ui_data.vcxproj
	scilab/modules/ui_data/src/jni/BrowseVar.cpp
	scilab/modules/ui_data/src/jni/BrowseVar.hxx

+++ scilab/tools/localization/updateLocalizationModule.sh
@@ -2,6 +2,7 @@
 # Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
 # Copyright (C) INRIA - 2007-2008 - Sylvestre Ledru
 # Copyright (C) DIGITEO - 2009-2011 - Sylvestre Ledru
+# Copyright (C) DIGITEO - 2011-2011 - Bruno JOFRET
 # This file must be used under the terms of the CeCILL.
 # This source file is licensed as described in the file COPYING, which
 # you should have received as part of this distribution.  The terms
@@ -11,9 +12,9 @@
 # This script goes into a module and updates the localization file by checking
 # the _( and gettext( calls in the code
 
-# TODO : 
+# TODO :
 # * Write Small documentation
-# 
+#
 
 if test $# -ne 1; then
     echo "This script goes into a module and updates the localization file "
@@ -28,7 +29,7 @@ if test $# -ne 1; then
 fi
 
 if test -z "$SCI"; then
-    echo "Please define the variable SCI" 
+    echo "Please define the variable SCI"
     exit 42
 fi
 
@@ -53,8 +54,8 @@ GUI_FILES="etc/*.xml"
 FAKE_C_FILE=scilab_fake_localization_file.c
 TIMEZONE="+0100"
 # Gettext arg
-XGETTEXT_OPTIONS="--add-location --strict --keyword=_ --from-code $FROM_CODE --omit-header --sort-output --no-wrap "
-    
+XGETTEXT_OPTIONS="--add-location --strict --keyword=_ --from-code $FROM_CODE --omit-header --no-wrap --sort-by-file"
+
 process_XML_files() {
 # First expression => remove line which does NOT contain label
 # Second expression =>  extract the content of the label and switch it to a gettext fake instruction
@@ -115,10 +116,10 @@ function process_module {
     if test $IS_MACROS -eq 1; then
         generate_find_command EXTENSIONS_MACROS
         local TARGETDIR=$TARGETDIR_MACROS
-        if test ! -d $TARGETDIR; then mkdir $TARGETDIR; fi
     else
         generate_find_command EXTENSIONS
     fi
+    if test ! -d $TARGETDIR; then mkdir $TARGETDIR; fi
 
     FILES=`eval $FILESCMD|tr "\n" " "`
 
@@ -131,6 +132,9 @@ function process_module {
     if test -n "$(ls $GUI_FILES 2>/dev/null)" -a $IS_MACROS -ne 1; then
         FILES="$FILES `ls $GUI_FILES`"
     fi
+
+    FILES=$(echo $FILES|sort)
+
     MODULE_NAME=`echo $MODULE|sed -e 's|./||'` # avoid to have ./module_name
 
     if test $IS_MACROS -eq 1; then
@@ -139,7 +143,7 @@ function process_module {
         echo "..... Parsing all sources in $PATHTOPROCESS"
     fi
 # Parse all the sources and get the string which should be localized
-    
+
 
     if test $IS_MACROS -eq 1; then
         MODULE_NAME=$MODULE_NAME-macros
@@ -158,17 +162,53 @@ function process_module {
     fi
 
     $XGETTEXT $XGETTEXT_OPTIONS -p $TARGETDIR/ -o $MODULE_NAME.pot.tmp $FILES > /dev/null
+    if test ! -f $MODULE_NAME.pot.tmp -a $IS_MACROS -eq 1; then
+        # Empty file => no string found
+        # We are modifing on the fly Scilab localization files
+        # 
+        # We need C strings format to be used as gettext key
+        # "" -> \"
+        # '' -> '
+        # '" -> \"
+        # "' -> ' -e "s/\"'/'/g" 
+        sed -i -e "s/\"\"/\\\"/g" -e "s/''/'/g" -e "s/'\"/\\\"/g" $TARGETDIR/$MODULE_NAME.pot.tmp
+        # We introduced invalid tag [msgstr "] and [msgid "]
+        # restore them [msgstr ""] and [msgid ""]
+        sed -i -e "s/msgstr \"$/msgstr \"\"/" -e "s/msgid \"$/msgid \"\"/" $TARGETDIR/$MODULE_NAME.pot.tmp
+    fi
+
     if test  -z "$CreationDate"; then
         # File not existing before ... Set the current date a POT-Creation-Date
         sed -e "s/MODULE/$MODULE_NAME/" -e "s/CREATION-DATE/`date +'%Y-%m-%d %H:%M'`$TIMEZONE/" -e "s/REVISION-DATE/`date +'%Y-%m-%d %H:%M'`$TIMEZONE/" $HEADER_TEMPLATE > $LOCALIZATION_FILE_US
     else
         sed -e "s/MODULE/$MODULE_NAME/" -e "s/CREATION-DATE/$CreationDate/" -e "s/REVISION-DATE/`date +'%Y-%m-%d %H:%M'`$TIMEZONE/" $HEADER_TEMPLATE > $LOCALIZATION_FILE_US
     fi
-    cat $LOCALIZATION_FILE_US.tmp >> $LOCALIZATION_FILE_US
+
+    msguniq -u $LOCALIZATION_FILE_US.tmp >> $LOCALIZATION_FILE_US 2> /dev/null
+
     rm $LOCALIZATION_FILE_US.tmp 2> /dev/null
+
+    MSGOUTPUT=$(msgcat $LOCALIZATION_FILE_US)
+    if test $? -ne 0; then
+        echo "Badly formated localization files"
+        exit 32
+    fi
     if test -z "$(msgcat $LOCALIZATION_FILE_US)"; then
         # empty template. Kill it!
         rm $LOCALIZATION_FILE_US
+    fi
+
+    if test $IS_MACROS -eq 1; then
+        LOCALIZATION_FILE_NATIVE=$(echo $LOCALIZATION_FILE_US|sed -e "s|-macros||g" -e "s|_macros||g")
+        if test ! -f $LOCALIZATION_FILE_NATIVE; then
+            # no native code. Copy the macro one
+            cp $LOCALIZATION_FILE_US $LOCALIZATION_FILE_NATIVE
+        else
+        # merge locale macros => native code
+            msgcat --use-first  -o $LOCALIZATION_FILE_NATIVE.tmp  $LOCALIZATION_FILE_NATIVE $LOCALIZATION_FILE_US
+            mv $LOCALIZATION_FILE_NATIVE.tmp $LOCALIZATION_FILE_NATIVE
+        fi
+        rm -rf $TARGETDIR_MACROS
     fi
 
     # Remove fake file used to extract string from XML
@@ -189,6 +229,6 @@ for MODULE in $MODULES; do
     cd $PATHTOPROCESS
     process_module "src"
     process_module "macros"
-
+    
     cd $SCI/
 done # Browse modules
